찐 수정파일/subway-mornitor/requirements.txt
Flask
requests
Flask-Cors
mysql-connector-python
